requirements.txt
pyquery
pyyaml
cymysql
